+++ Dau.Core/Domain/Catalog/DormitoryBlockTranslation.cs
@@ -7,7 +7,7 @@ namespace Dau.Core.Domain.Catalog
 {
    public class DormitoryBlockTranslation:  BaseLanguage
     {
-        public int DormitoryBlockNonTransId { get; set; }
+        public long DormitoryBlockNonTransId { get; set; }
         public string Name { get; set; }
 
        public string Description { get; set; }
