+++ Dau.Core/Domain/ContentManagement/MessageTemplateTranslation.cs
@@ -9,7 +9,7 @@ namespace Dau.Core.Domain.ContentManagement
     {
      
       
-       public int MessageTemplateNonTransId { get; set; }
+       public long MessageTemplateNonTransId { get; set; }
       public string Subject { get; set; }
 
        public string Body { get; set; }
