+++ Dau.Core/Domain/Facility/FacilityOptionTranslation.cs
@@ -6,7 +6,7 @@ namespace Dau.Core.Domain.Facility
 {
     public partial class FacilityOptionTranslation : BaseLanguage
     {
-        public int FacilityOptionNonTransId { get; set; }
+        public long FacilityOptionNonTransId { get; set; }
        
         public string FacilityOption { get; set; }
         public string FacilityOptionDescription { get; set; }
